chore: code fixes
