+++ components/Content/Post.tsx
@@ -1,7 +1,7 @@
 import Image from 'next/image';
 import Link from 'next/link';
 import { useMemo } from 'react';
-import { Article } from './News';
+import { Article } from 'components/Article';
 import { formatDate } from 'utils/formatDate';
 
 export const Post = ({
